requirements.txt
brotli==1.2.0
cffi==2.0.0
cssselect2==0.8.0
fonttools==4.61.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
pillow==12.0.0
pycparser==2.23
pydyf==0.12.1
pyphen==0.17.2
PyQt6==6.10.0
PyQt6-Qt6==6.10.1
PyQt6_sip==13.10.2
python-docx==1.2.0
qrcode==8.2
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
